Remove staff required from all product management views to superuser required setting

+++ products/views.py
@@ -86,15 +86,19 @@ def product_management(request):
     return render(request, 'products/product_management.html', context)
 
 
-@staff_member_required
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+        
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
             product = form.save()
             messages.success(request, 'Successfully added product!')
-            return redirect(reverse('product_detail', args=[product.id]))
+            return redirect(reverse('products:product_detail', args=[product.id]))
         else:
             messages.error(
                 request,
@@ -110,16 +114,20 @@ def add_product(request):
     return render(request, template, context)
 
 
-@staff_member_required
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
         if form.is_valid():
             form.save()
             messages.success(request, 'Successfully updated product!')
-            return redirect(reverse('product_detail', args=[product.id]))
+            return redirect(reverse('products:product_detail', args=[product.id]))
         else:
             messages.error(
                 request,
@@ -137,13 +145,17 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
-@staff_member_required
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
-    return redirect(reverse('products'))
+    return redirect(reverse('products:products'))
 
 
 @login_required
